+++ libs/qsynedit/qsynedit/document.h
@@ -712,28 +712,18 @@ public:
 
     bool fullUndoImposible() const;
 
-    int maxMemoryUsage() const;
-    void setMaxMemoryUsage(int newMaxMemoryUsage);
-
 signals:
     void addedUndo();
 protected:
-    void ensureMaxEntries();
     bool inBlock();
     unsigned int getNextChangeNumber();
-    void addMemoryUsage(PUndoItem item);
-    void reduceMemoryUsage(PUndoItem item);
 protected:
     size_t mBlockChangeNumber;
     int mBlockLock;
-    int mBlockCount; // count of action blocks;
-    int mMemoryUsage;
     size_t mLastPoppedItemChangeNumber;
     size_t mLastRestoredItemChangeNumber;
     bool mFullUndoImposible;
     QVector<PUndoItem> mItems;
-    int mMaxUndoActions;
-    int mMaxMemoryUsage;
     unsigned int mNextChangeNumber;
     unsigned int mInitialChangeNumber;
     bool mInsideRedo;